Move argparse to config module.
